common: Split pybtp/btp.py into multiple files
because it has already grown too much.
