+++ pybtp/btp/__init__.py
@@ -1,0 +1,24 @@
+#
+# auto-pts - The Bluetooth PTS Automation Framework
+#
+# Copyright (c) 2017, Intel Corporation.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms and conditions of the GNU General Public License,
+# version 2, as published by the Free Software Foundation.
+#
+# This program is distributed in the hope it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+
+"""BTP"""
+
+from pybtp.btp.btp import *
+from pybtp.btp.gap import *
+from pybtp.btp.gatt import *
+from pybtp.btp.l2cap import *
+from pybtp.btp.mesh import *
+from pybtp.btp.mmdl import *
+from pybtp.btp.sm import *
